backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
PyJWT[crypto]
requests
